requirements.txt
langchain
langchain-openai
langchain-community
langchain-mcp-adapters
chromadb
python-dotenv
mcp
anyio

